requirements.txt

numpy==1.16.0
pytest==4.2.0
biopython==1.72
scipy==1.1.0
scikit-learn==0.19.2
matplotlib==2.2.3
pandas==0.23.4
seaborn==0.9.0
